requirements.txt
anthropic>=0.25.0
openai>=1.0.0
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
pysqlite3-binary>=0.5.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pypdf>=4.0.0
tiktoken>=0.6.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.29.0
plotly>=5.0.0
python-dateutil>=2.9.0
